+++ src/@types/services.ts
@@ -1,4 +1,4 @@
-import { ServiceType } from '@/constants/serviceType'
+import { ServiceType } from '@/constants'
 
 export type ServiceType = typeof ServiceType
 export type ServiceTypes = ServiceType[keyof ServiceType]
