+++ src/constants/content.ts
@@ -1,0 +1,1 @@
+export const PER_PAGE = 10
